Add autoattack as package
